Ch1: added exercises before sqrt, and an answer to 1.6
I decided to go back and include answers/explanations to earlier exercises.
Where code isn't asked for, I've included the example code and an answer in
the comments.

I've also included an implementation of sqrt that uses cond rather than if
to demonstrate that cond is a special form and evaluates the same way that
if does.  This is specified in the text, but I wrote a code example to
affirm my understanding.