Moved ch1-sqrt.scm to sicp-1.1.scm to align with book/site layout.
As I just added answers and code for exercises other than the square root
problem, the file needed to be renamed to reflect its contents.  From now
on I'll go with this format as it lines up with the answers on the
following site, so I'm assuming it should work for the rest of the book.

http://community.schemewiki.org/?SICP-Solutions